requirements.txt

google-generativeai
python-dotenv
.streamlit
secrets.toml